refactor(types): общий базовый класс для lifecycle-инвалидации индексов
SonarQube Quality Gate падал по дублированию (>3% на новом коде): четыре
@EventListener-метода дословно совпадали в InferredVariableTypeIndex и
CallStatementByReceiverIndex (~20 строк × 2). Вынес их в package-private
AbstractDocumentLifecycleClearableIndex с abstract clear(URI); Spring наследует
@EventListener в конкретных подклассах-бинах, поэтому индексы просто его
расширяют и реализуют clear(URI). Дубль убран.

Заодно тест на агрегацию: byReceiver возвращает оба callStatement'а для одного
ресивера (hasSize(2)) — ревью-замечание CodeRabbit.
